+++ resources/views/link-archive.blade.php
@@ -4,6 +4,21 @@
 <div class="container">
     <div class="row justify-content-center">
         <div class="col-12 col-md-8">
+            @if($link->hasArchive())
+                <div class="card mb-3">
+                    <div class="card-header">{{ __('Current archive') }}</div>
+
+                    <div class="card-body">
+                        <form action="{{ route('link.archive-delete', $link->id) }}" method="POST">
+                            @csrf
+
+                            <a href="{{ route('link.archive-download', [$link->id, csrf_token()]) }}" class="btn btn-sm btn-primary">{{ __('Download archive') }}</a>
+                            <button type="submit" class="btn btn-sm btn-danger" name="type" value="pdf">{{ __('Delete archive') }}</button>
+                        </form>
+                    </div>
+                </div>
+            @endif
+
             <div class="card">
                 <div class="card-header">{{ __('New archive') }}</div>
 
@@ -21,7 +36,7 @@
                             <table class="table table-condensed table-bordered">
                                 @if(in_array('media', $providers))
                                     <tr>
-                                        <td>{{ __('Archive as Media (audio or video)') }}</td>
+                                        <td>{{ __('Archive media') }}</td>
                                         <td class="text-right">
                                             <button type="submit" class="btn btn-sm btn-primary btn-block" name="type" value="media">{{ __('Choose') }}</button>
                                         </td>
@@ -30,7 +45,7 @@
 
                                 @if(in_array('pdf', $providers))
                                     <tr>
-                                        <td>{{ __('Archive as PDF') }}</td>
+                                        <td>{{ __('Archive PDF') }}</td>
                                         <td class="text-right">
                                             <button type="submit" class="btn btn-sm btn-primary btn-block" name="type" value="pdf">{{ __('Choose') }}</button>
                                         </td>
@@ -41,21 +56,6 @@
                     @endif
                 </div>
             </div>
-
-            @if($link->hasArchive())
-            <div class="card mt-3">
-                <div class="card-header">{{ __('Current archive') }}</div>
-
-                <div class="card-body">
-                    <form action="{{ route('link.archive-delete', $link->id) }}" method="POST">
-                        @csrf
-
-                        <a href="{{ route('link.archive-download', [$link->id, csrf_token()]) }}" class="btn btn-sm btn-primary">{{ __('Download archive') }}</a>
-                        <button type="submit" class="btn btn-sm btn-danger" name="type" value="pdf">{{ __('Delete archive') }}</button>
-                    </form>
-                </div>
-            </div>
-            @endif
         </div>
     </div>
 </div>