+++ app/Chest.php
@@ -3,7 +3,6 @@
 namespace App;
 
 use App\Concerns\Models\Postable;
-use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 
 class Chest extends Model
@@ -19,19 +18,9 @@ class Chest extends Model
         'content' => 'json',
     ];
 
-    public function getHashIdAttribute(): string
-    {
-        return hashid_encode($this->id);
-    }
-
     public function getPermalinkAttribute(): string
     {
         return route('chest.view', $this->hash_id);
-    }
-
-    public function scopeHashIdIs(Builder $query, string $hash): Builder
-    {
-        return $query->where('id', hashid_decode($hash));
     }
 
     public function toSearchableArray()