+++ resources/views/layouts/partials/head.blade.php
@@ -6,7 +6,6 @@
     <meta name="api-token" content="{{ auth()->user()->api_token }}">
 @endif
 @stack('meta')
-@include('feed::links')
 <link rel="dns-prefetch" href="//fonts.gstatic.com">
 <link href="https://fonts.googleapis.com/css?family=Nunito" rel="stylesheet">
 <link href="{{ mix('css/app.css') }}" rel="stylesheet">