+++ routes/web.php
@@ -8,8 +8,6 @@ Route::auth([
     'verify' => false,
 ]);
 
-Route::feeds();
-
 Route::get('/', 'BrowseController@index')->name('home');
 Route::get('/tag/{tag}', 'BrowseController@tag')->name('tag');