+++ changelog.md
@@ -1,10 +1,19 @@
-# Unreleased
+# 1.2.0
 
 ## Added
 
 - Loader on LinkForm, StoryForm and ChestForm
 - Add original URL to card-link
 - Ability to "Save and archive" a link
+
+## Changed
+
+- Laravel 6.0
+- Simplified Hashid
+
+## Deleted
+
+- Feeds 
 
 # 1.1.1